[issue37] Remove unused functions and move to place used.
